# Speed up import of pyomo.environ