--- a/pyomo/common/dependencies.py
+++ b/pyomo/common/dependencies.py
@@ -2,12 +2,13 @@
 #
 #  Pyomo: Python Optimization Modeling Objects
 #  Copyright 2017 National Technology and Engineering Solutions of Sandia, LLC
-#  Under the terms of Contract DE-NA0003525 with National Technology and 
-#  Engineering Solutions of Sandia, LLC, the U.S. Government retains certain 
+#  Under the terms of Contract DE-NA0003525 with National Technology and
+#  Engineering Solutions of Sandia, LLC, the U.S. Government retains certain
 #  rights in this software.
 #  This software is distributed under the 3-clause BSD License.
 #  ___________________________________________________________________________
 
+import inspect
 import importlib
 import logging
 
@@ -37,6 +38,86 @@ class ModuleUnavailable(object):
         logging.getLogger(logger).warning(
             self._error_message_)
 
+class DeferredImportModule(object):
+    """Dummy object that serves as a module placeholder until the first time
+    getattr is called, at which point it imports the module and returns
+    the module attribute.
+
+    This object is returned by attempt_import() in liu of the module in
+    the case that the module import fails.  Any attempts to access
+    attributes on this object will raise a DeferredImportError
+    exception.
+    """
+    def __init__(self, indicator):
+        self._indicator_flag = indicator
+
+    def __getattr__(self, attr):
+        self._indicator_flag.resolve()
+        return getattr(self._indicator_flag._module, attr)
+
+
+class DeferredImportIndicator(object):
+    """Placeholder indicating if an import was successful.
+
+    This object serves as a placeholder for the Boolean indicator if a
+    deferred module import was successful.  Casting this instance to
+    bool will cause the import to be attempted.  The actual import logic
+    is here and not in the DeferredImportModule to reduce the number of
+    attributes on the DeferredImportModule.
+    """
+
+    def __init__(self, name, alt_names, error_message, only_catch_importerror,
+                 minimum_version, original_globals, callback):
+        self._names = (name,)
+        if alt_names:
+            self._names += tuple(alt_names)
+        self._error_message = error_message
+        self._only_catch_importerror = only_catch_importerror
+        self._minimum_version = minimum_version
+        self._original_globals = original_globals
+        self._callback = callback
+        self._module = None
+        self._available = None
+
+    def resolve(self):
+        if self._module is None:
+            # Only attempt the import once
+            self._module, self._available = attempt_import(
+                name=self._names[0],
+                error_message=self._error_message,
+                only_catch_importerror=self._only_catch_importerror,
+                minimum_version=self._minimum_version,
+                defer_check=False,
+                callback=self._callback,
+            )
+            # Replace myself in the original globals() where I was
+            # declared
+            self.replace_self_in_globals(self._original_globals)
+
+        # Replace myself in the caller globals (to avoid calls to
+        # this method in the future)
+        _globals = inspect.currentframe().f_back.f_back.f_globals
+        self.replace_self_in_globals(_globals)
+
+    def replace_self_in_globals(self, _globals):
+        for name in self._names:
+            if ( name in _globals
+                 and isinstance(_globals[name], DeferredImportModule)
+                 and _globals[name]._indicator_flag is self ):
+                _globals[name] = self._module
+            for flag_name in (name
+
+    + '_available', 'has_' + name):
+                if flag_name in _globals and _globals[flag_name] is self:
+                    _globals[flag_name] = self._available
+
+    def __nonzero__(self):
+        self.resolve()
+        return self._available
+
+    __bool__ = __nonzero__
+
+
 try:
     from packaging import version as _version
     _parser = _version.parse
@@ -47,11 +128,16 @@ except ImportError:
     from pkg_resources import parse_version as _parser
 
 def _check_version(module, min_version):
-    return _parser(min_version) <= _parser(module.__version__)
-    
+    version = getattr(module
+
+    , '__version__', '0')
+    return _parser(min_version) <= _parser(version)
+
 
 def attempt_import(name, error_message=None, only_catch_importerror=True,
-                   minimum_version=None):
+                   minimum_version=None, alt_names=None, callback=None,
+                   defer_check=True):
+
     """Attempt to import the specified module.
 
     This will attempt to import the specified module, returning a
@@ -65,42 +151,91 @@ def attempt_import(name, error_message=None, only_catch_importerror=True,
     .. doctest::
 
        >>> from pyomo.common.dependencies import attempt_import
-       >>> np, numpy_available = attempt_import('numpy')
+       >>> numpy, numpy_available = attempt_import('numpy')
+
+    The import can be "deferred" until the first time the code either
+    attempts to access the module or checks the boolean value of the
+    available flag.  This allows optional dependencies to be declared at
+    the module scope but not imported until they are actually used by
+    the module (thereby speeding up the initial package import).
+    Deferred imports are handled by two helper classes
+    (DeferredImportModule and DeferredImportIndicator).  Upon actual
+    import, DeferredImportIndicator.resolve() attempts to replace those
+    objects (in both the local and original global namespaces) with the
+    imported module and boolean flag so that subsequent uses of the
+    module do not incur any overhead due to the delayed import.
 
     Parameters
     ----------
-    name: `str`
+    name: str
         The name of the module to import
 
-    error_message: `str`, optional
+    error_message: str, optional
         The message for the exception raised by ModuleUnavailable
 
-    only_catch_importerror: `bool`, optional
-        If True, exceptions other than ImportError raised during module
-        import will be reraised.  If False, any exception will result in
-        returning a ModuleUnavailable object.
+    only_catch_importerror: bool, optional
+        If True (the default), exceptions other than ImportError raised
+        during module import will be reraised.  If False, any exception
+        will result in returning a ModuleUnavailable object.
+
+    minimum_version: str, optional
+        The minimum acceptable module version (retrieved from
+        module.__version__)
+
+    alt_names: list, optional
+        A list of common alternate names by which to look for this
+        module in the globals() namespaces.  For example, the alt_names
+        for NumPy would be ['np']
+
+    callback: function, optional
+        A function with the signature "`fcn(module, available)`" that
+        will be called after the import is first attempted.
+
+    defer_check: bool, optional
+        If True (the default), then the attempted import is deferred
+        until the first use of either the module or the availability
+        flag.  The method will return instances of DeferredImportModule
+        and DeferredImportIndicator.
 
     Returns
     -------
     : module
-        the imported module or an instance of :py:class:`ModuleUnavailable`
+        the imported module, or an instance of
+        :py:class:`ModuleUnavailable`, or an instance of
+        :py:class:`DeferredImportModule`
     : bool
-        Boolean indicating if the module import succeeded
+        Boolean indicating if the module import succeeded or an instance
+        of "py:class:`DeferredImportIndicator`
+
     """
+    # If we are going to defer the check until later, return the
+    # deferred import module object
+    if defer_check:
+        indicator = DeferredImportIndicator(
+            name=name,
+            alt_names=alt_names,
+            error_message=error_message,
+            only_catch_importerror=only_catch_importerror,
+            minimum_version=minimum_version,
+            original_globals=inspect.currentframe().f_back.f_globals,
+            callback=callback)
+        return DeferredImportModule(indicator), indicator
+
     try:
         module = importlib.import_module(name)
-        if minimum_version is None:
-            return module, True
-        elif _check_version(module, minimum_version):
+        if minimum_version is None or _check_version(module, minimum_version):
+            if callback is not None:
+                callback(module, True)
             return module, True
         elif error_message:
+            version = getattr(module, '__version__', 'UNKNOWN')
             error_message += " (version %s does not satisfy the minimum " \
-                             "version %s)" % (
-                                 module.__version__, minimum_version)
+                             "version %s)" % (version, minimum_version)
         else:
+            version = getattr(module, '__version__', 'UNKNOWN')
             error_message = "The %s module version %s does not satisfy " \
                             "the minimum version %s" % (
-                                name, module.__version__.minimum_version)
+                                name, version, minimum_version)
     except ImportError:
         pass
     except:
@@ -110,17 +245,41 @@ def attempt_import(name, error_message=None, only_catch_importerror=True,
     if not error_message:
         error_message = "The %s module (an optional Pyomo dependency) " \
                         "failed to import" % (name,)
-    return ModuleUnavailable(error_message), False
+
+    module = ModuleUnavailable(error_message)
+    if callback is not None:
+        callback(module, False)
+    return module, False
 
 #
 # Common optional dependencies used throughout Pyomo
 #
 
-yaml, yaml_available = attempt_import('yaml')
-if yaml_available and hasattr(yaml, 'SafeLoader'):
-    yaml_load_args = {'Loader': yaml.SafeLoader}
-else:
-    yaml_load_args = {}
+yaml_load_args = {}
+def _yaml_importer_callback(module, available):
+    # Recent versions of PyYAML issue warnings if the Loader argument is
+    # not set
+    if available and hasattr(module, 'SafeLoader'):
+        yaml_load_args['Loader'] = module.SafeLoader
 
-numpy, numpy_available = attempt_import('numpy')
-scipy, scipy_available = attempt_import('scipy')
+def _finalize_scipy(module, available):
+    if not available:
+        return
+    # Import key subpackages that we will want to assume are present
+    import scipy.sparse
+    import scipy.spatial
+
+def _finalize_pympler(module, available):
+    if not available:
+        return
+    # Import key subpackages that we will want to assume are present
+    import pympler.muppy
+
+yaml, yaml_available = attempt_import(
+    'yaml', callback=_yaml_importer_callback)
+pympler, pympler_available = attempt_import(
+    'pympler', callback=_finalize_pympler)
+numpy, numpy_available = attempt_import('numpy', alt_names=['np'])
+scipy, scipy_available = attempt_import('scipy', callback=_finalize_scipy)
+networkx, networkx_available = attempt_import('networkx', alt_names=['nx'])
+casadi, casadi_available = attempt_import('casadi')