… pyomo/common/tests/dep_mod.py
… pyomo/common/tests/dep_mod_except.py
… pyomo/common/tests/test_dependencies.py
… pyomo/contrib/fbbt/tests/test_fbbt.py
… pyomo/contrib/fbbt/tests/test_interval.py
… pyomo/contrib/parmest/parmest.py
… pyomo/contrib/pynumero/__init__.py
… pyomo/contrib/trustregion/tests/TestTRConfig.py
… pyomo/core/base/PyomoModel.py
… pyomo/core/base/config.py
… pyomo/core/base/units_container.py
… pyomo/core/expr/sympy_tools.py
… pyomo/core/kernel/matrix_constraint.py
… pyomo/core/kernel/piecewise_library/util.py
… pyomo/core/plugins/transform/add_slack_vars.py
… pyomo/core/tests/examples/test_kernel_examples.py
… pyomo/core/tests/unit/kernel/test_parameter.py
… pyomo/core/tests/unit/test_block.py
… pyomo/core/tests/unit/test_component.py
… pyomo/core/tests/unit/test_set.py
… pyomo/core/tests/unit/test_units.py
… pyomo/dae/plugins/colloc.py
… pyomo/dae/simulator.py
… pyomo/dae/tests/test_colloc.py
… pyomo/dae/tests/test_simulator.py
… pyomo/dataportal/TableData.py
… pyomo/dataportal/plugins/db_table.py
… pyomo/dataportal/plugins/sheet.py
… pyomo/dataportal/plugins/xml_table.py
… pyomo/dataportal/tests/param4.baseline.csv