@blnicho 2020-03-16
> This PR reduces import pyomo.environ from ~4s to roughly 0.5s.

Do these times factor in the changes you've also made to PyUtilib to defer imports?