@jsiirola 2020-03-16
Yes: the net import time reduction includes this PR plus a bit of work in PyUtilib.  The changes to pyutilib were mostly to avoid importing things like `openpyxl`, and were necessary to avoid importing `yaml` (which was imported in about a dozen places spread across both projects).