@jsiirola 2020-03-17
I believe this is ready to review.  The only "failing" check is coverage, and it's hard to tell from the codecov report where the coverage really decreased.