@blnicho 2020-03-18
@jsiirola I merged the ParmEst PR before I realized there were conflicts with this one. Once we resolve the conflicts I think this is ready to merge.